Tests: passing local pytest; enable GH Actions CI; add io module and docs autosummary; replace RTD config
