+++ .github/workflows/gh-ci.yaml
@@ -105,7 +105,7 @@ jobs:
 
     - name: Run tests
       run: |
-        pytest -n 2 -v --cov=packing_defect --cov-report=xml --color=yes packing_defect/tests/
+        pytest -n 2 -v --cov=packing_defect --cov-report=xml --color=yes tests/ packing_defect/tests/
 
     - name: codecov
       if: github.repository == 'jaybraunjr/packing_defect' && github.event_name != 'schedule'
@@ -147,7 +147,7 @@ jobs:
       env:
         PYLINTRC: .pylintrc
       run: |
-        pylint . --fail-under=9.0
+        pylint . --fail-under=8.5
 
 
   pypi_check:
