+++ .readthedocs.yaml
@@ -1,14 +1,12 @@
-# readthedocs.yaml
-
 version: 2
 
 build:
   os: ubuntu-22.04
   tools:
-    python: "mambaforge-4.10"
+    python: "mambaforge-22.9"
 
 sphinx:
-  configuration: docs/source/conf.py   
+  configuration: docs/source/conf.py
 
 python:
   install:
@@ -17,3 +15,4 @@ python:
 
 conda:
   environment: docs/requirements.yaml
+
