+++ docs/source/autosummary/packing_defect.core.analyzers.base.rst
@@ -4,15 +4,6 @@ packing\_defect.core.analyzers.base
 .. automodule:: packing_defect.core.analyzers.base
 
    
-   
-   
-
-   
-   
-   
-
-   
-   
    .. rubric:: Classes
 
    .. autosummary::
@@ -20,11 +11,3 @@ packing\_defect.core.analyzers.base
       AnalysisResult
       BaseDefectAnalyzer
    
-   
-
-   
-   
-   
-
-
-
