+++ docs/source/autosummary/packing_defect.core.analyzers.packing.rst
@@ -4,26 +4,9 @@ packing\_defect.core.analyzers.packing
 .. automodule:: packing_defect.core.analyzers.packing
 
    
-   
-   
-
-   
-   
-   
-
-   
-   
    .. rubric:: Classes
 
    .. autosummary::
    
       PackingDefectAnalyzer
    
-   
-
-   
-   
-   
-
-
-
