+++ docs/source/autosummary/packing_defect.core.analyzers.radius.rst
@@ -4,26 +4,9 @@ packing\_defect.core.analyzers.radius
 .. automodule:: packing_defect.core.analyzers.radius
 
    
-   
-   
-
-   
-   
-   
-
-   
-   
    .. rubric:: Classes
 
    .. autosummary::
    
       RadiusDefectAnalyzer
    
-   
-
-   
-   
-   
-
-
-
