+++ docs/source/autosummary/packing_defect.core.cluster.rst
@@ -4,26 +4,9 @@ packing\_defect.core.cluster
 .. automodule:: packing_defect.core.cluster
 
    
-   
-   
-
-   
-   
-   
-
-   
-   
    .. rubric:: Classes
 
    .. autosummary::
    
       DefectClustering
    
-   
-
-   
-   
-   
-
-
-
