+++ docs/source/autosummary/packing_defect.core.interactions.rst
@@ -4,11 +4,6 @@ packing\_defect.core.interactions
 .. automodule:: packing_defect.core.interactions
 
    
-   
-   
-
-   
-   
    .. rubric:: Functions
 
    .. autosummary::
@@ -16,21 +11,9 @@ packing\_defect.core.interactions
       process_multiple
       run
    
-   
-
-   
-   
    .. rubric:: Classes
 
    .. autosummary::
    
       ProteinDefectInteractions
    
-   
-
-   
-   
-   
-
-
-
