+++ docs/source/autosummary/packing_defect.core.topology.rst
@@ -4,26 +4,9 @@ packing\_defect.core.topology
 .. automodule:: packing_defect.core.topology
 
    
-   
-   
-
-   
-   
-   
-
-   
-   
    .. rubric:: Classes
 
    .. autosummary::
    
       TopologyReader
    
-   
-
-   
-   
-   
-
-
-
