+++ docs/source/autosummary/packing_defect.core.vis.rst
@@ -4,27 +4,9 @@ packing\_defect.core.vis
 .. automodule:: packing_defect.core.vis
 
    
-   
-   
-
-   
-   
    .. rubric:: Functions
 
    .. autosummary::
    
       main
-      plot_defect_data
    
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
