+++ docs/source/autosummary/packing_defect.core.visualization.rst
@@ -4,14 +4,11 @@ packing\_defect.core.visualization
 .. automodule:: packing_defect.core.visualization
 
    
-   
-   
+   .. rubric:: Functions
 
+   .. autosummary::
    
-   
-   
-
-   
+      plot_defect_data
    
    .. rubric:: Classes
 
@@ -19,11 +16,3 @@ packing\_defect.core.visualization
    
       LeafletVisualizer
    
-   
-
-   
-   
-   
-
-
-
