+++ docs/source/autosummary/packing_defect.core.writer.rst
@@ -4,26 +4,9 @@ packing\_defect.core.writer
 .. automodule:: packing_defect.core.writer
 
    
-   
-   
-
-   
-   
    .. rubric:: Functions
 
    .. autosummary::
    
       write_defect_coordinates
    
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
