+++ docs/source/autosummary/packing_defect.io.rst
@@ -1,0 +1,13 @@
+packing\_defect.io
+==================
+
+.. automodule:: packing_defect.io
+
+   
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   writer
