+++ docs/source/autosummary/packing_defect.io.writer.rst
@@ -1,0 +1,12 @@
+packing\_defect.io.writer
+=========================
+
+.. automodule:: packing_defect.io.writer
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      write_defect_coordinates
+
