+++ docs/source/autosummary/packing_defect.run_defect.rst
@@ -4,26 +4,9 @@ packing\_defect.run\_defect
 .. automodule:: packing_defect.run_defect
 
    
-   
-   
-
-   
-   
    .. rubric:: Functions
 
    .. autosummary::
    
       build_packing_analyzer
    
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
