+++ docs/source/autosummary/packing_defect.run_radius.rst
@@ -4,26 +4,9 @@ packing\_defect.run\_radius
 .. automodule:: packing_defect.run_radius
 
    
-   
-   
-
-   
-   
    .. rubric:: Functions
 
    .. autosummary::
    
       build_radius_analyzer
    
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
