+++ docs/source/autosummary/packing_defect.run_utils.rst
@@ -4,26 +4,9 @@ packing\_defect.run\_utils
 .. automodule:: packing_defect.run_utils
 
    
-   
-   
-
-   
-   
    .. rubric:: Functions
 
    .. autosummary::
    
       run_analysis
    
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
