+++ packing_defect/core/writer.py
@@ -1,10 +1,17 @@
-# packing_defect/core/writer.py
+"""Deprecated shim for writer functions.
 
-import os
-import json
-from packing_defect.utils import write_combined_gro
+Re-exports write_defect_coordinates from packing_defect.io.writer for backward compatibility,
+and exposes write_combined_gro from utils for existing monkeypatch targets in tests/code.
+"""
+
+from packing_defect.utils import write_combined_gro  # noqa: F401
+
 
 def write_defect_coordinates(protein, defect_atoms, dims, path):
+    import os
     os.makedirs(os.path.dirname(path), exist_ok=True)
     write_combined_gro(protein, defect_atoms, dims, path)
+
+
+__all__ = ["write_defect_coordinates", "write_combined_gro"]
