+++ packing_defect/io/__init__.py
@@ -1,0 +1,2 @@
+"""I/O helpers for packing_defect."""
+
