+++ packing_defect/io/writer.py
@@ -1,0 +1,8 @@
+import os
+from packing_defect.utils import write_combined_gro
+
+
+def write_defect_coordinates(protein, defect_atoms, dims, path):
+    os.makedirs(os.path.dirname(path), exist_ok=True)
+    write_combined_gro(protein, defect_atoms, dims, path)
+
