+++ packing_defect/tests/test_packing.py
@@ -1,7 +1,4 @@
 
-import sys, os
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-
 # tests/test_packing_analyzer.py
 
 import os
